% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmappr.R
\name{collapse_heatmap}
\alias{collapse_heatmap}
\title{Collapse a heatmap if the x and y-axes include subgroups}
\usage{
collapse_heatmap(p, collapse_fn = mean, ...)
}
\arguments{
\item{p}{a ggplot heatmap with x, y, and fill aesthetics defined}

\item{collapse_fn}{the function to use to combine all the inner values}

\item{...}{additional arguments to collapse_fn}
}
\description{
If the factor defined on the x-axis is a sample but there are actually
replicates for each sample, then defining the plot with aes(x=sample)
and then collapse_heatmap yields the mean of each replicate.
}
\details{
Similarly, for the y-axis, if you're interested in higher-order taxa, set
aes(y=Genus) or similar and then collapse_heatmap to yield the mean abundance
for that taxa.

If both axes contain subgroups, this averages over all of them.
}

