% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmappr.R
\name{heatmap_scale_1}
\alias{heatmap_scale_1}
\title{Heatmap color scale 1}
\usage{
heatmap_scale_1(..., na.value = "white", threshold = 0.9)
}
\arguments{
\item{...}{arguments passed to scale_fill_gradientn}

\item{threshold}{the limit where the high-range color appears}
}
\description{
A color scale that moves from light blue to purple for most of the range,
with a high-range (default, >90%) color of bright fuschia.
Drop in replacement for scale_fill_gradientn.
}

