% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmappr.R
\name{make_square}
\alias{make_square}
\title{make_square}
\usage{
make_square(p)
}
\arguments{
\item{p}{a ggplot object, ideally with discrete x and y aesthetics set
(i.e. a heatmap)}
}
\value{
the same object, with the aspect ratio fixed to be the (number of
rows)/(number of columns)
}
\description{
make_square
}

